Move helpers/users_helper.rb to controllers/concerns/eos_account.rb to use other controllers
